// sourcecode/BTrak.Packages/snova-dashboard-module/Dashboard-module-components/dashboard-components/src/lib/snovasys-dashboard-management/models/project.model.ts
export class Project {
  projectResponsiblePersonId: string;
  projectId: string;
  projectTypeId: string;
  isArchived: boolean;
  projectName: string;
  projectTypeName: string;
  projectStatusColor: string;
  userName: string;
  createdDateTime: Date;
  fullName: string;
  roleName: string;
  profileImage: string;
  projectResponsiblePerson: any;
  activeGoalCount: number;
  numberOfReds: number;
  timeStamp: any;
  projectLabel: string;
  testSuiteCount: number;
  testRunCount: number;
  milestoneCount: number;
  totalCount: number;
  reportCount: number;
  casesCount: number;
  isDateTimeConfiguration: boolean;
  isSprintsConfiguration: boolean;
  viewGoalsPermissionCount: number;
}